+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # 0.1.0
-- Initial version
+- Added ModelAttribute and ModelConvention
 
 Full release notes can be found at: https://github.com/micro-elements/MicroElements.Design.Annotations/blob/master/CHANGELOG.md